Relocated module locations of some functions
