+++ galsampler/host_halo_selection.py
@@ -3,11 +3,8 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
-from astropy.table import Table
-from halotools.utils import crossmatch
 
 from .cython_kernels import source_halo_index_selection_kernel
-from .host_halo_binning import halo_bin_indices
 
 
 def source_halo_index_selection(first, last, num_select):
@@ -47,39 +44,3 @@ def source_halo_index_selection(first, last, num_select):
     return np.array(source_halo_index_selection_kernel(first, last, num_select))
 
 
-def compute_richness(host_halo_ids):
-    """
-    """
-    nhalos = host_halo_ids.shape[0]
-    host_halo_ngals = np.zeros(nhalos).astype('i4')
-
-    uval, idx_uval, counts = np.unique(host_halo_ids, return_index=True, return_counts=True)
-    host_halo_ngals[idx_uval] = counts
-
-    idxA, idxB = crossmatch(host_halo_ids, host_halo_ids[idx_uval])
-    host_halo_ngals[idxA] = host_halo_ngals[idx_uval][idxB]
-
-    return host_halo_ngals
-
-
-def format_source_catalog(source_halos, halo_property_bins):
-    """
-    Examples
-    --------
-    >>> from galsampler.tests import fake_source_galaxy_catalog
-    >>> source_halos = fake_source_galaxy_catalog()
-    >>> num_bins_mass, num_bins_conc = 12, 11
-    >>> mass_bins = np.logspace(10, 15, num_bins_mass)
-    >>> conc_bins = np.logspace(1.5, 20, num_bins_conc)
-    >>> halo_property_bins = dict(host_halo_mass=mass_bins, host_halo_conc=conc_bins)
-    >>> formatted_source_halos = format_source_catalog(source_halos, halo_property_bins)
-    """
-    source_halos = Table(source_halos)
-    source_halos['host_halo_ngals'] = compute_richness(source_halos['host_halo_id'])
-
-    source_halos['halo_bin_number'] = halo_bin_indices(
-        {key: source_halos[key] for key in halo_property_bins.keys()}, halo_property_bins)
-
-    source_halos.sort(('halo_bin_number', 'host_halo_id', 'satellite'))
-
-    return source_halos
